+++ graph/adapter-output-spring-data-neo4j-sdn6/src/main/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/SpringDataNeo4JResearchProblemAdapter.kt
@@ -138,13 +138,6 @@ class SpringDataNeo4JResearchProblemAdapter(
     ): Page<ResearchProblemRepository.ContributorPerProblem> =
         neo4jRepository.findContributorsLeaderboardPerProblem(problemId.toResourceId(), pageable)
 
-    override fun findAuthorsLeaderboardPerProblem(
-        problemId: ThingId,
-        pageable: Pageable
-    ): Page<ResearchProblemRepository.AuthorPerProblem> =
-        neo4jRepository.findAuthorsLeaderboardPerProblem(problemId.toResourceId(), pageable)
-            .map { it.toAuthorPerProblem() }
-
     override fun findResearchProblemForDataset(datasetId: ThingId): Iterable<Resource> =
         neo4jRepository.findResearchProblemForDataset(datasetId.toResourceId()).map { it.toResource() }